Change minimum versions of Node and eslint in integration tests (#83)
Node v16 is out of support. So we test against Node v18 and v20.

eslint 7 was released 4 years ago so its fair to not actively test again
eslint 6. Speculativiely adding eslint 9 tests but not sure if that will
work

Also fix testing of multiple eslint versions

We weren't doing custom installations of different eslint versions
so we were always only testing one eslint version!

Added eslint-v9 test directory but it fails right now so not adding it
to CI.